backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
google-genai==1.9.0
python-dotenv==1.0.0
nest-asyncio==1.5.8